refactor: rename BackButton for reset email
